rename style
