+++ ui/src/pages/ResetPassword/ResetPassword.jsx
@@ -15,7 +15,7 @@ import IconVisibility from '@mui/icons-material/Visibility'
 import IconVisibilityOff from '@mui/icons-material/VisibilityOff'
 
 // STYLES
-import useLayoutStyles from 'styles/layoutAuthenticationHalf'
+import useLayoutStyles from 'styles/layoutAuthentication'
 
 const ResetPassword = () => {
   const layoutClasses = useLayoutStyles()
