+++ ui/src/pages/SignUp/SignUp.jsx
@@ -18,7 +18,7 @@ import IconVisibility from '@mui/icons-material/Visibility'
 import IconVisibilityOff from '@mui/icons-material/VisibilityOff'
 
 // STYLES
-import useLayoutStyles from 'styles/layoutAuthenticationHalf'
+import useLayoutStyles from 'styles/layoutAuthentication'
 
 const SignUp = () => {
   const layoutClasses = useLayoutStyles()
