+++ ui/src/styles/layoutAuthentication.js
